src/FC.CodeFlix.Catalog.Domain/Validation/DomainValidation.cs
using FC.CodeFlix.Catalog.Domain.Exceptions;

namespace FC.CodeFlix.Catalog.Domain.Validation;
public class DomainValidation
{
    public static void NotNull(object? target, string fieldName)
    {
        if (target is null)
            throw new EntityValidationException($"{fieldName} is required");
    }

    public static void NotNullOrEmpty(string? target, string fieldName)
    {
        if (string.IsNullOrWhiteSpace(target))
            throw new EntityValidationException($"{fieldName} is required");
    }

    public static void MinLength(string target, int minLength, string fieldName)
    {
        if(target.Length < minLength)
            throw new EntityValidationException($"{fieldName} must be at least {minLength} characters");
    }

    public static void MaxLength(string target, int maxLength, string fieldName)
    {
        if (target.Length > maxLength)
            throw new EntityValidationException($"{fieldName} must be less or equal {maxLength} characters");
    }
}
